refactor: BridgeController를 BridgeGame으로 변경
