+++ src/main/java/bridge/Application.java
@@ -1,14 +1,13 @@
 package bridge;
 
-import bridge.controller.BridgeController;
+import bridge.controller.BridgeGame;
 import bridge.view.InputView;
 import bridge.view.OutputView;
 
 public class Application {
 
     public static void main(String[] args) {
-        BridgeController bridgeController = new BridgeController();
-
-        bridgeController.run();
+        BridgeGame bridgeGame = new BridgeGame(new InputView(), new OutputView());
+        bridgeGame.run();
     }
 }
